frontend_admin/requirements.txt
streamlit
requests
pandas # Для удобного отображения табличных данных
python-dotenv # Для загрузки URL бэкенда из .env